fix bug 2

+++ helpers/splitLocation.ts
@@ -43,7 +43,7 @@ export const splitFunction = ({
 
     let shiftPoint: ShiftPointCrops = {
       axis: parseInt(axis),
-      direction: parseInt(direction),
+      direction: parseFloat(direction),
       lon: parseFloat(point.split(",")[0].trim()),
       lat: parseFloat(point.split(",")[1].trim()),
       z: parseFloat(point.split(",")[2].trim()),
